requirements.txt
requests>=2.25.0
asyncio>=3.4.3
colorama>=0.4.4
PyYAML>=6.0
markdown>=3.3.0
pyflakes>=3.0.0
tabulate>=0.9.0
rich>=13.0.0